Fixed requests to the game indexer.Added requests to generate events

+++ Assets/Scripts/Near/Models/Tokens/Players/FieldPlayer/FieldPlayer.cs
@@ -7,7 +7,7 @@ namespace Near.Models.Tokens.Players.FieldPlayer
     {
         [JsonIgnore]
         public FieldPlayerStats Stats { get; set; }
-        public int teamwork { get; set; }
+        public float teamwork { get; set; }
         public string native_position { get; set; }
 
         public int  number_of_penalty_events { get; set; }
